+++ CHANGELOG.md
@@ -18,11 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.4] - Unreleased
 
+- Fixed hardware security settings decoding
+- Fixed power supply structure minimum length
 - Fixed cooling device structure minimum length
 - Fixed current probe structure minimum length
 - Fixed temperature probe structure mimimum length
 - Fixed voltage probe structure minimum length
-- Fixed hardware security settings decoding
 - Fixed baseboard structure minumum length
 
 ## [0.1.3] - January 11, 2026